requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
biliard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-3-jet==1.0.8
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-redis==5.0.0
django-rest-knox==4.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==9.2.0
idna==3.2
kombu==5.1.0
oauthlib==3.1.1
Pillow==8.3.2
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
